requirements.txt
flask==3.1.0
apscheduler==3.10.4
requests==2.32.3
pyyaml==6.0.2